removed warnings

+++ projects/MOM2/pr_meslep.cpp
@@ -386,7 +386,7 @@ void perens_t::average_r_pr_meslep(perens_t &out,const array<double,2> w) const
 	      const size_t i=im_r_im_r_meslepins_iop_iproj_imeslepmom_ind(im_r_im_r_iop_iproj_imeslepmom_comp);
 	      pr_rave[out_i]+=w[r]*pr[i];
 	      
-	      for(int ijack=0;ijack<=njacks;ijack++)
+	      for(size_t ijack=0;ijack<=njacks;ijack++)
 		cout<<" "<<ijack<<" "<<im_r_im_r_meslepins_iop_iproj_imeslepmom_ind.descr(i)<<" "<<pr[i][ijack]<<endl;
 	    }
 	  pr_rave[out_i]/=nr;
@@ -501,10 +501,10 @@ void perens_t::val_chir_extrap_pr_meslep(perens_t &out) const
 	    for(size_t ipow=0;ipow<coeffs[meslepins].size();ipow++)
 	      {
 		cout<<"ipow: "<<ipow<<endl;
-		for(int ijack=0;ijack<=njacks;ijack++)
+		for(size_t ijack=0;ijack<=njacks;ijack++)
 		  cout<<" "<<ijack<<" "<<r_r_iop_iproj_imeslepmom_ind.descr(r_r_iop_iproj_imeslepmom)<<" "<<coeffs[meslepins][ipow][ijack]<<endl;
 	      }
-		    
+	    
 	    if(std::isnan(coeffs[meslepins][0][0])) coeffs[meslepins]=0.0;
 	    
 	    const size_t iout=out.im_r_im_r_meslepins_iop_iproj_imeslepmom_ind({0,r_in,0,r_ou,meslepins,iop,iproj,imeslepmom});
@@ -514,10 +514,10 @@ void perens_t::val_chir_extrap_pr_meslep(perens_t &out) const
 		auto xminmax=minmax_element(x.begin(),x.end());
 		double xmin=*xminmax.first*0.9;
 		double xmax=*xminmax.second*1.1;
-		for(int ijack=0;ijack<=njacks;ijack++)
+		for(size_t ijack=0;ijack<=njacks;ijack++)
 		  {
 		    cout<<" ijack: "<<ijack<<endl;
-		    for(int ipoint=0;ipoint<y_plot.size();ipoint++)
+		    for(size_t ipoint=0;ipoint<y_plot.size();ipoint++)
 		      cout<<" "<<x[ipoint]<<" "<<y_plot[ipoint][ijack]<<" "<<y_plot[ipoint].err()<<endl;
 		  }
